Renamed job for generating user skill rates for user and updated specs
