--- app/jobs/base_skill_rates_generator_job.rb
@@ -1,7 +1,0 @@
-class BaseSkillRatesGeneratorJob
-  include SuckerPunch::Job
-
-  def perform(user_id:)
-    Skills::UserSkillRatesGenerator.new(user_id: user_id).call
-  end
-end
