+++ app/jobs/create_rates_for_user_job.rb
@@ -1,0 +1,7 @@
+class CreateRatesForUserJob
+  include SuckerPunch::Job
+
+  def perform(user_id:)
+    Skills::UserSkillRatesGenerator.new.generate_all_for_user(user_id: user_id)
+  end
+end
